change parameter of mauve from (float,int,float,int) to (float,float,float,float) min lcb len and max lcb len are set to percentage of the shorter sequence, not a int of length any more. also the order of those is different. now it's (quality of lcb, min cover, min lcb length, max lcb length) since we now have default value of all four parameters after mauve, I modify the part under nexus, fill in the default (35.0,0.35,0.01,0.10) if input is missing.

+++ src/methods.ml
@@ -138,7 +138,7 @@ type transform_cost_matrix = [
 type median_solver_chosen = [ `MGR | `SimpleLK | `ChainedLK | `COALESTSP | `BBTSP |
 `Siepel | `Albert | `Vinh  ]
 
-type annotate_tool = [ `Mauve of (float*int*float*int) | `Default of (int*int*int) ]
+type annotate_tool = [ `Mauve of (float*float*float*float) | `Default of (int*int*int) ]
 
 type align_meth = [ `NewKK | `Default ]
